kaem.run: New file.
* kaem.run: New file.

+++ lib/m2/ntoab.c
@@ -25,7 +25,7 @@
 char *__itoa_buf;
 
 char *
-ntoab (int x, int base, int signed_p)
+ntoab (long x, int base, int signed_p)
 {
   if (__itoa_buf == 0)
     __itoa_buf = malloc (20);
